Assessments
Project contains 3 files, 1 html, 1 javascript and 1 css file.

This contains a single html form to upload an image file and show display it. We can add description of any area in the image by clicking that area. The description and the co-ordinates are also listed in a table and a red dot marking is done at that particular poing in the image also. whenever the user hover the red-dot, the description for that particular area is shown.